Heavy Refactoring
* Entirely changed into smol
* Commands, Handles, Mutexes are all async now
* Project structure refactored
