+++ .runer
@@ -74,11 +74,11 @@ flows:
       tasks:
         - { id: 0, type: Blueprint, name: db, job: container }
         # - { id: 1, type: Blueprint, name: test, job: shell, depends: 0 }
-        # - { id: 1, type: Env, name: dev, job: set, depends: 0 }
-        # - { id: 2, type: Blueprint, name: api, job: image, depends: 1 }
-        # - { id: 3, type: Blueprint, name: api, job: container, depends: 2 }
-        # - { id: 4, type: Blueprint, name: crawler, job: image, depends: 1 }
-        # - { id: 5, type: Blueprint, name: crawler, job: container, depends: 4 }
+        - { id: 1, type: Env, name: dev, job: set, depends: 0 }
+        - { id: 2, type: Blueprint, name: api, job: image, depends: 1 }
+        - { id: 3, type: Blueprint, name: api, job: container, depends: 2 }
+        - { id: 4, type: Blueprint, name: crawler, job: image, depends: 1 }
+        - { id: 5, type: Blueprint, name: crawler, job: container, depends: 4 }
       pkg_dependencies: ["psql", "sqlx"]
         
 env:
